+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+Added
+  * AZ-573: Allow Vnet in separate resource group
+
 Changed
   * AZ-572: Revamp examples and improve CI